Create GKE spin-up and spin-down scripts
This commit adds two new scripts, gke-spin-up.sh and gke-spin-down.sh, which are used to create and delete a GKE cluster respectively. The gke-spin-up.sh script creates a GKE cluster with preemptible nodes, while the gke-spin-down.sh script deletes the GKE cluster. These scripts are useful for managing the lifecycle of the GKE cluster in an automated manner.

Signed-off-by: Ihor Dvoretskyi <ihor@linux.com>